refactor: integrate eth_api
