+++ src/lib.rs
@@ -62,7 +62,6 @@ mod message_pool;
 mod metrics;
 mod networks;
 mod rpc;
-mod rpc_api;
 mod rpc_client;
 mod shim;
 mod state_manager;
