+++ src/rpc/mod.rs
@@ -8,7 +8,7 @@ pub mod beacon_api;
 pub mod chain_api;
 mod channel;
 pub mod common_api;
-mod eth_api;
+pub mod eth_api;
 pub mod gas_api;
 pub mod mpool_api;
 pub mod net_api;
@@ -33,7 +33,6 @@ use crate::rpc::auth_layer::AuthLayer;
 use crate::rpc::channel::RpcModule as FilRpcModule;
 pub use crate::rpc::channel::CANCEL_METHOD_NAME;
 use crate::rpc::state_api::*;
-use crate::rpc_api::eth_api::*;
 
 use fvm_ipld_blockstore::Blockstore;
 use hyper::server::conn::AddrStream;
